feat(website): [T002149] website-db.ts Stage 1 split — extract Customer/Bug/SiteSettings/Vacation/Legal to website-core-db.ts
Extracted five domains from website-db.ts (1939→1518 lines, -421 lines):
- Customer (upsertCustomer, listPendingEnrollments, declineEnrollment,
  getCustomerFullById, getCustomerByKeycloakId, setCustomerNumber,
  setAdminNumber, setIsAdmin)
- Bug Tickets (insertBugTicket, resolveBugTicket, archiveBugTicket,
  getBugTicketStatus, getBugTicketWithComments, appendBugTicketComment,
  reopenBugTicket)
- Site Settings (initSiteSettingsTable, getSiteSetting, setSiteSetting)
- Vacation/Blackout (getVacationPeriods, saveVacationPeriods)
- Legal Pages (initLegalPagesTable, getLegalPage, saveLegalPage)

New file website-core-db.ts (473 lines) keeps all extracted logic with
its own imports (pool, initTicketsSchema, transitionTicket).
website-db.ts re-exports everything for backward compatibility.

Unused imports removed: transitionTicket, Customer type (direct).
Internal imports added: upsertCustomer (for assignMeeting),
BugTicketRow (for listBugTickets).

Verification:
- tsc --noEmit: passed (no errors)
- vitest run: 350 files, 2777 tests passed
- madge --circular: no cycles found (website/src/lib/ level)

Not in scope: Stage 2 (T002150).
